don't count partial string matches as equal
Previously, typing e.g. "cda" at the prompt would count as a match for
"cd", since streq() wasn't checking to ensure the length of the cstr
wasn't longer than that of the &str (it was, however, checking if it was
shorter).

This solves that issue by comparing the length at the end of the
equality test.