closes #8: rejoin split blocks when empty
The old allocator wasn't looking at the block next to a freed block when
deallocating memory. Instead it was just freeing the block it was on.
This inspects the buddy of a given block when it is freed, and if the
buddy is also empty then it goes up and changes the parent from SPLIT to
UNUSED.

This solves the problem of constantly using more memory even though
memory was still being freed, something you could pick up on earlier if
you added debug statements to allocator.rs that print the locations
of allocs when they happen.